moved twitter handle functionality into a function
