+++ plugin-hrm-form/src/components/search/SearchForm/index.jsx
@@ -25,6 +25,7 @@ import { namespace, configurationBase, searchContactsBase } from '../../../state
 import { getNumberFromTask } from '../../../services/ContactService';
 import { localizedSource } from '../../PreviousContactsBanner';
 import { channelTypes } from '../../../states/DomainConstants';
+import {formatNumberFromTask} from '../../../utils/formatters'
 
 const getField = value => ({
   value,
@@ -127,8 +128,7 @@ class SearchForm extends Component {
     const { task } = this.props;
 
     const contactNumberFromTask = getNumberFromTask(task);
-    const checkBoxName =
-      task.channelType === channelTypes.twitter ? `@${task.attributes.twitterUserHandle}` : contactNumberFromTask;
+    const checkBoxName = formatNumberFromTask(task, contactNumberFromTask)
 
     const handleChangePreviousContactsCheckbox = () => {
       const value = contactNumber === '' ? contactNumberFromTask : '';
